simplify tsconfig setup

+++ scripts/clean.mjs
@@ -1,11 +1,15 @@
 import * as Glob from "glob"
 import * as Fs from "node:fs"
 
-const dirs = [".", ...Glob.sync("packages/*/", ...Glob.sync("packages/ai/*/"))]
+const dirs = [".", ...Glob.sync("packages/*/")]
 dirs.forEach((pkg) => {
-  const files = [".tsbuildinfo", "build", "dist", "coverage"]
+  const files = [".tsbuildinfo", "tsconfig.tsbuildinfo", "docs", "build", "dist", "coverage"]
 
   files.forEach((file) => {
+    if (pkg === "." && file === "docs") {
+      return
+    }
+
     Fs.rmSync(`${pkg}/${file}`, { recursive: true, force: true }, () => {})
   })
 })
